backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
PyYAML==6.0.2
pytest==8.4.1
pytest-asyncio==0.24.0
